+++ custom_components/wollm/__init__.py
@@ -12,6 +12,7 @@ from .services import async_register_services, async_unregister_services
 
 PLATFORMS: tuple[Platform, ...] = (
     Platform.BUTTON,
+    Platform.NUMBER,
     Platform.SELECT,
     Platform.SENSOR,
     Platform.SWITCH,